apps/telegram_template/requirements.txt
jq
python-telegram-bot